Security Flaw Fixed
Don’t allow a user to delete the “first” variable flag out of the
users.json database. They should only be able to delete the users out
of the database.

Typo

Fixed type

Fixed Flaw

Prevent users from adding a username called “first” because a variable
named “first” is stored in the users.json database. Would not want them
to overwrite it!

Removed empty file

+++ login.js
@@ -9,20 +9,25 @@ var genRandomString = function(length){
 };
 
 var addUser = function(user,pass) {
-	const username = user;
-	const password = pass;
-	const salt_gen = genRandomString(64);
-	const hash_gen = crypto.pbkdf2Sync(password, salt_gen+username, 1000, 64, `sha512`).toString(`hex`);
+	if(user != "first"){
+		const username = user;
+		const password = pass;
+		const salt_gen = genRandomString(64);
+		const hash_gen = crypto.pbkdf2Sync(password, salt_gen+username, 1000, 64, `sha512`).toString(`hex`);
 	
-	db.set(username,{salt:salt_gen,hash:hash_gen});
-	if(db.get('first')){
-		db.set('first',false);
+		db.set(username,{salt:salt_gen,hash:hash_gen});
+		if(db.get('first')){
+			db.set('first',false);
+		}
+  		db.sync();
 	}
-  	db.sync();
 }
 
 var getUser = function(user){
-  return db.get(user);
+  if(user != "first"){
+  	return db.get(user);
+  }
+  	return false;
 }
 
 var isFirst = function(){
@@ -30,7 +35,7 @@ var isFirst = function(){
 }
 
 var deleteUser = function(user){
-	if(db.has(user)){
+	if(db.has(user) && user != "first"){
 		db.delete(user);	
 	}
 }
